[LCR-238] refactor: 오늘 입장자 수 조회 기능 테스트 코드 리팩토링

+++ src/main/java/com/lgcns/domain/entrance/dto/response/DailyEntrantCountResponse.java
@@ -3,8 +3,8 @@ package com.lgcns.domain.entrance.dto.response;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 public record DailyEntrantCountResponse(
-        @Schema(description = "입장자 수", example = "100") Integer entrantCount) {
-    public static DailyEntrantCountResponse of(Integer entrantCount) {
+        @Schema(description = "입장자 수", example = "100") Long entrantCount) {
+    public static DailyEntrantCountResponse of(Long entrantCount) {
         return new DailyEntrantCountResponse(entrantCount);
     }
 }
